Move Allowed* to own namespace
